+++ packages/messaging/src/events/event/payloads.ts
@@ -1,50 +1,35 @@
-import type { EventType as ContractEventType, EventState } from '@volontariapp/contracts';
-import type { EventLocation } from '@volontariapp/shared';
+import type { IEventIdPayload, IUserIdPayload } from '../index.js';
 
 export enum EventEventMessagingType {
-  EVENT_CHANGED = 'event.changed',
   EVENT_CREATED = 'event.created',
   EVENT_DELETED = 'event.deleted',
-  REQUIREMENT_CHANGED = 'requirement.changed',
-  TAG_CHANGED = 'tag.changed',
-  EVENT_TAG_LINKED = 'event.tag_linked',
+
+  EVENT_GEOCODED = 'event.geocoded',
+  EVENT_GEOCODING_FAILED = 'event.geocoding_failed',
+
+  EVENT_CREATION_SUCCESSFULL = 'event.creation_successfull', // WS -> [EVENT] -> SAGA DONE
+  EVENT_CREATION_FAILED = 'event.creation_failed', // WS -> [EVENT, SOCIAL] -> SAGA FAILED
+  EVENT_DELETION_SUCCESSFULL = 'event.deletion_successfull', // WS -> [EVENT] -> DELETE EVENT
+  EVENT_DELETION_FAILED = 'event.deletion_failed', // WS -> TODO: Reflechir
 }
 
-export interface IEventPayload {
-  id: string;
-  name: string;
-  description: string;
-  startAt: Date;
-  endAt: Date;
-  type: ContractEventType;
-  state: EventState;
-  awardedImpactScore: number;
-  maxParticipants: number;
-  organizerId: string;
+export interface IEventCreatedPayload extends IEventIdPayload, Partial<IUserIdPayload> {
   localisationName: string;
-  eventLocation: EventLocation;
-  createdAt: Date;
-  updatedAt: Date;
 }
+export interface IEventDeletedPayload extends IEventIdPayload, Partial<IUserIdPayload> {}
 
-export interface IRequirementPayload {
-  id: string;
-  name: string;
-  description?: string;
-  quantity: number;
-  currentQuantity: number;
-  isSystem: boolean;
-  createdBy?: string;
-}
+export interface IEventGeocodedPayload extends IEventIdPayload, Partial<IUserIdPayload> {}
 
-export interface ITagPayload {
-  id: string;
-  name: string;
-  slug: string;
-  balise: string;
-}
+export interface IEventGeocodingFailedPayload extends IEventIdPayload, Partial<IUserIdPayload> {}
 
-export interface IEventTagLinkedPayload {
-  eventsId: string;
-  tagsId: string;
-}
+export interface IEventCreationSuccessfullPayload
+  extends IEventIdPayload,
+    Partial<IUserIdPayload> {}
+
+export interface IEventCreationFailedPayload extends IEventIdPayload, Partial<IUserIdPayload> {}
+
+export interface IEventDeletionSuccessfullPayload
+  extends IEventIdPayload,
+    Partial<IUserIdPayload> {}
+
+export interface IEventDeletionFailedPayload extends IEventIdPayload, Partial<IUserIdPayload> {}