+++ packages/shared/src/enums/streams.enum.ts
@@ -10,9 +10,11 @@ export enum UserStream {
 export enum EventStream {
   EVENT_JOB_OUTBOX_SUCCESS = 'event:job:outbox:success',
   EVENT_JOB_OUTBOX_FAILURE = 'event:job:outbox:failure',
-  EVENT_EVENTS = 'event:events',
+  EVENT_CREATED = 'event:created',
+  EVENT_DELETED = 'event:deleted',
   EVENT_TAGS = 'event:tags',
   EVENT_REQUIREMENTS = 'event:requirements',
+  EVENT_SUCCESSFULLY_CREATED = 'event:successfully_created',
 }
 
 export enum PostStream {
@@ -40,6 +42,8 @@ export enum WebsocketStream {
   WS_BADGE = 'ws:badge',
   WS_POST_CREATED_FEEDBACK = 'ws:post-created-feedback',
   WS_POST_DELETED_FEEDBACK = 'ws:post-deleted-feedback',
+  WS_EVENT_CREATED_FEEDBACK = 'ws:event-created-feedback',
+  WS_EVENT_DELETED_FEEDBACK = 'ws:event-deleted-feedback',
 }
 
 export const Streams = {